refactor: move wildcard replacement stages
