+++ easyuse_anima/wildcard/expansion.py
@@ -6,8 +6,9 @@ import bisect
 import hashlib
 import math
 import re
-from collections.abc import Callable, Iterable
+from collections.abc import Callable, Iterable, Sequence
 from dataclasses import dataclass
+from typing import Protocol
 
 from . import sources as _wildcard_sources
 from .models import WildcardExpansionBudget, WildcardOption
@@ -85,6 +86,10 @@ def _parse_count_spec(spec: str, selector: _Selector) -> int | None:
     minimum = int(left) if left else 0
     maximum = int(right) if right else minimum
     return selector.count_from_range(minimum, maximum)
+
+
+class _WildcardOptionLookup(Protocol):
+    def options_for(self, raw_key: str) -> Sequence[WildcardOption]: ...
 
 
 def _utf8_width(char: str) -> int:
@@ -328,6 +333,120 @@ class _ExpansionState:
             cursor = match.end()
         output.extend(current.slice_segments(cursor, len(source)))
         return _ExpansionText(output)
+
+
+def _expand_multiselect_options(
+    options: list[WildcardOption],
+    library: _WildcardOptionLookup,
+) -> tuple[Sequence[WildcardOption], str | None]:
+    if len(options) != 1:
+        return options, None
+    match = WILDCARD_FULL_RE.match(options[0].text.strip())
+    if not match:
+        return options, None
+    raw_key = match.group("keyword")
+    return (
+        library.options_for(raw_key),
+        _wildcard_sources._normalize_wildcard_key(raw_key),
+    )
+
+
+def _replace_dynamic(
+    current: _ExpansionText,
+    state: _ExpansionState,
+    selector: _Selector,
+    library: _WildcardOptionLookup,
+) -> _ExpansionText:
+    def replace(
+        match: re.Match[str],
+        key_stack: tuple[str, ...],
+    ) -> _Replacement | None:
+        body = match.group(1)
+        raw_options = _split_unescaped(body, "|")
+        if not raw_options:
+            return None
+        first_parts = raw_options[0].split("$$")
+        if len(first_parts) > 1:
+            count = _parse_count_spec(first_parts[0], selector)
+            if count is None:
+                return None
+            separator = ", "
+            if len(first_parts) == 2:
+                first_candidate = first_parts[1]
+            else:
+                separator = first_parts[1]
+                first_candidate = "$$".join(first_parts[2:])
+            candidate_text = "|".join([first_candidate, *raw_options[1:]])
+            options, expanded_key = _expand_multiselect_options(
+                _parse_dynamic_options(candidate_text),
+                library,
+            )
+            selected = selector.choose_many(options, count)
+            if not selected:
+                return None
+            candidate_stack = key_stack + (
+                (expanded_key,) if expanded_key is not None else ()
+            )
+            return state.candidate(
+                (option.text for option in selected),
+                separator,
+                candidate_stack,
+            )
+
+        options = _parse_dynamic_options(body)
+        selected = selector.choose_one(options)
+        if selected is None:
+            return None
+        return state.candidate((selected.text,), "", key_stack)
+
+    return state.replace_matches(current, DYNAMIC_RE, replace)
+
+
+def _replace_quantified_wildcards(
+    current: _ExpansionText,
+    state: _ExpansionState,
+    selector: _Selector,
+    library: _WildcardOptionLookup,
+) -> _ExpansionText:
+    def replace(
+        match: re.Match[str],
+        key_stack: tuple[str, ...],
+    ) -> _Replacement | None:
+        count = max(0, int(match.group("quantifier")))
+        raw_key = match.group("keyword")
+        options = library.options_for(raw_key)
+        selected = selector.choose_many(options, count)
+        key = _wildcard_sources._normalize_wildcard_key(raw_key)
+        if not selected or key is None:
+            return None
+        return state.candidate(
+            (option.text for option in selected),
+            ", ",
+            key_stack + (key,),
+        )
+
+    return state.replace_matches(current, WILDCARD_QUANTIFIER_RE, replace)
+
+
+def _replace_file_wildcards(
+    current: _ExpansionText,
+    state: _ExpansionState,
+    selector: _Selector,
+    library: _WildcardOptionLookup,
+) -> _ExpansionText:
+    def replace(
+        match: re.Match[str],
+        key_stack: tuple[str, ...],
+    ) -> _Replacement | None:
+        raw_key = match.group("keyword")
+        options = library.options_for(raw_key)
+        selected = selector.choose_one(options)
+        key = _wildcard_sources._normalize_wildcard_key(raw_key)
+        if selected is None or key is None:
+            return None
+        return state.candidate((selected.text,), "", key_stack + (key,))
+
+    return state.replace_matches(current, WILDCARD_RE, replace)
 
 
 def has_wildcard_syntax(text: str) -> bool:
